Working on nn input feature extraction executed on all 56 anims for 750 frames JFrame updated to hold all 8 local features (feetDist excluded from serialization) need to run feature extraction again

+++ Assets/ObjectsForJson/JAnimation.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Text;
+using UnityEngine;
 
 namespace ObjectsForJson
 {
@@ -37,7 +38,8 @@ namespace ObjectsForJson
             List<float> peaks = new List<float>(20);
 
             // taglio i valori in eccesso 
-            frames.RemoveRange(750, frames.Count - 750);
+            if (frames.Count > 750)
+                frames.RemoveRange(750, frames.Count - 750);
